moved from of the codec specific parsing into their own files
