+++ ftl_app/main.h
@@ -56,6 +56,3 @@
 void charon_install_ctrlc_handler();
 void charon_loop_until_ctrlc();
 BOOL ctrlc_pressed();
-int gettimeofday(struct timeval * tp, struct timezone * tzp);
-int timeval_subtract(struct timeval *result, struct timeval *x, struct timeval *y);
-float timeval_to_ms(struct timeval *tv);
